+++ docs/CHANGELOG.rst
@@ -20,6 +20,7 @@ Added
 Changed
 -------
 - Change ``UploadConfig`` type to ``APIView``
+- Emit auditlog in background task
 
 Changed
 -------